Remove holiday cache persistence
