+++ src/main/java/com/company/officecommute/web/ApiConvertor.java
@@ -1,22 +1,15 @@
 package com.company.officecommute.web;
 
-import com.company.officecommute.domain.overtime.Holiday;
 import com.company.officecommute.domain.overtime.HolidayResponse;
-import com.company.officecommute.domain.overtime.HolidaySyncStatus;
 import com.company.officecommute.global.exception.HolidayDataUnavailableException;
-import com.company.officecommute.repository.overtime.HolidayRepository;
-import com.company.officecommute.repository.overtime.HolidaySyncStatusRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
 
-import java.time.Clock;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.YearMonth;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
@@ -34,25 +27,15 @@ public class ApiConvertor {
 
     private final RestTemplate restTemplate;
     private final ApiProperties apiProperties;
-    private final HolidayRepository holidayRepository;
-    private final HolidaySyncStatusRepository holidaySyncStatusRepository;
-    private final Clock clock;
 
     public ApiConvertor(
             RestTemplate restTemplate,
-            ApiProperties apiProperties,
-            HolidayRepository holidayRepository,
-            HolidaySyncStatusRepository holidaySyncStatusRepository,
-            Clock clock
+            ApiProperties apiProperties
     ) {
         this.restTemplate = restTemplate;
         this.apiProperties = apiProperties;
-        this.holidayRepository = holidayRepository;
-        this.holidaySyncStatusRepository = holidaySyncStatusRepository;
-        this.clock = clock;
     }
 
-    @Transactional
     public long countNumberOfStandardWorkingDays(YearMonth yearMonth) {
         Set<LocalDate> holidays = getHolidays(yearMonth);
         int lengthOfMonth = yearMonth.lengthOfMonth();
@@ -63,28 +46,10 @@ public class ApiConvertor {
         return numberOfWeekDays - numberOfHolidays;
     }
 
-    /**
-     * 공휴일 데이터를 API에서 가져와 DB에 저장합니다.
-     * 관리자 수동 동기화 API에서 호출하여 캐시를 갱신합니다.
-     */
-    @Transactional
-    public void refreshHolidays(YearMonth yearMonth) {
-        try {
-            List<HolidayResponse.Item> items = fetchHolidaysFromApi(yearMonth);
-            Set<LocalDate> holidays = convertToLocalDate(items);
-            saveHolidaysToDatabase(yearMonth, holidays);
-            log.info("공휴일 갱신 성공: {}-{}", yearMonth.getYear(), yearMonth.getMonthValue());
-        } catch (Exception e) {
-            log.warn("공휴일 갱신 실패. 기존 캐시 유지: {}-{}, 오류: {}",
-                    yearMonth.getYear(), yearMonth.getMonthValue(), e.getMessage());
-        }
-    }
-
     private Set<LocalDate> getHolidays(YearMonth yearMonth) {
         try {
             List<HolidayResponse.Item> items = fetchHolidaysFromApi(yearMonth);
             Set<LocalDate> holidays = convertToLocalDate(items);
-            saveHolidaysToDatabase(yearMonth, holidays);
             log.info("공휴일 API 호출 성공: {}-{}", yearMonth.getYear(), yearMonth.getMonthValue());
             return holidays;
         } catch (Exception e) {
@@ -92,23 +57,6 @@ public class ApiConvertor {
                     yearMonth, e.getMessage(), e);
             throw new HolidayDataUnavailableException(HOLIDAY_DATA_UNAVAILABLE_MESSAGE);
         }
-    }
-
-    private void saveHolidaysToDatabase(YearMonth yearMonth, Set<LocalDate> holidays) {
-        int year = yearMonth.getYear();
-        int month = yearMonth.getMonthValue();
-
-        holidayRepository.deleteByYearAndMonth(year, month);
-
-        List<Holiday> holidayEntities = holidays.stream()
-                .map(date -> new Holiday(year, month, date))
-                .toList();
-        holidayRepository.saveAll(holidayEntities);
-
-        HolidaySyncStatus syncStatus = holidaySyncStatusRepository.findByYearAndMonth(year, month)
-                .orElseGet(() -> new HolidaySyncStatus(year, month, LocalDateTime.now(clock)));
-        syncStatus.markSyncedAt(LocalDateTime.now(clock));
-        holidaySyncStatusRepository.save(syncStatus);
     }
 
     private List<HolidayResponse.Item> fetchHolidaysFromApi(YearMonth yearMonth) {
