+++ src/main/resources/db/migration/V6__drop_holiday_cache_tables.sql
@@ -1,0 +1,2 @@
+DROP TABLE IF EXISTS holiday_sync_status;
+DROP TABLE IF EXISTS holiday;
